# frontend/main/openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateOut": {
        "description": "One extracted reading, as a reviewer needs to see it before naming it.",
        "properties": {
          "candidate_id": {
            "format": "uuid",
            "title": "Candidate Id",
            "type": "string"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The extractor confidence as recorded; it is not a semantic-type confidence.",
            "title": "Confidence"
          },
          "corroboration_lane": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corroboration Lane"
          },
          "corroboration_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corroboration Status"
          },
          "crop_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Storage key of the crop of this reading's region, when one was cut.",
            "title": "Crop Key"
          },
          "page_index": {
            "title": "Page Index",
            "type": "integer"
          },
          "raw_text": {
            "title": "Raw Text",
            "type": "string"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The reading as exact text, `25 1/2 in`. Null when the token carried no unit and was recorded without a value, which is most text on a drawing.",
            "title": "Value"
          }
        },
        "required": [
          "candidate_id",
          "page_index",
          "raw_text"
        ],
        "title": "CandidateOut",
        "type": "object"
      },
      "CandidatesOut": {
        "description": "Everything read for this package that nobody has yet said the meaning of.",
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/CandidateOut"
            },
            "title": "Candidates",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "candidates",
          "total"
        ],
        "title": "CandidatesOut",
        "type": "object"
      },
...
      "ConfirmIn": {
        "description": "A reviewer naming what a reading is.",
        "properties": {
          "semantic_type": {
            "description": "A member of the rulebook's vocabulary, e.g. `CT010`. Free text is refused.",
            "title": "Semantic Type",
            "type": "string"
          }
        },
        "required": [
          "semantic_type"
        ],
        "title": "ConfirmIn",
        "type": "object"
      },
...
      "ConfirmedOut": {
        "description": "The evidence a confirmation produced.",
        "properties": {
          "canonical_observation_id": {
            "format": "uuid",
            "title": "Canonical Observation Id",
            "type": "string"
          },
          "semantic_type": {
            "title": "Semantic Type",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "canonical_observation_id",
          "semantic_type",
          "status"
        ],
        "title": "ConfirmedOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI-assisted shop drawing review. The AI reads; deterministic Python decides; a reviewer signs off.",
    "title": "Graniti Vicentia V1",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/packages/{package_id}/candidates": {
      "get": {
        "description": "Every reading of this package that carries a value, with its crop.\n\n**Only readings that carry a value.** A token with no unit was recorded without one \u2014 deliberately,\nbecause a bare `38` is a dimension whose unit is unknown \u2014 and there is nothing for a reviewer to\nconfirm about it. They are still in the database, and a reviewer who wants to see what was thrown\naway is asking a different question than this one.\n\nOrdered by page and then by when it was read, so the list follows the drawing rather than the\ndatabase's insertion order, and two loads put the same reading in the same place.",
        "operationId": "list_candidates_api_v1_projects__project_id__packages__package_id__candidates_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "package_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Package Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidatesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "What the extractor read, waiting for somebody to say what it is",
        "tags": [
          "confirmations"
        ]
      }
    },
    "/api/v1/projects/{project_id}/packages/{package_id}/candidates/{candidate_id}/confirm": {
      "post": {
        "description": "Turn one reading into evidence a rule may be run against.\n\nThe value is not in the request, and that is the point of the whole story: the reviewer is\nconfirming the number the extractor read, not supplying one. A body carrying a value would be the\nold form path wearing a new name.\n\nRuns in the request rather than a background task because it is one row's worth of work and a\nreviewer is waiting on the answer \u2014 and because the audit event naming them has to commit with it.",
        "operationId": "confirm_candidate_api_v1_projects__project_id__packages__package_id__candidates__candidate_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "package_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Package Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Say what an extracted reading is, so the rules can use it",
        "tags": [
          "confirmations"
        ]
      }
    },
    "/api/v1/projects/{project_id}/packages/{package_id}/candidates/{candidate_id}/crop": {
      "get": {
        "description": "Return the stored pixels a reviewer must inspect before naming a proposal.\n\nThis is deliberately candidate-scoped: the reading is still untyped, so it must not be exposed\nas a finding or redline.  The SQL path proves both the candidate and its crop belong to the\npackage's current revision, then the stored digest is checked before bytes leave the service.",
        "operationId": "candidate_crop_api_v1_projects__project_id__packages__package_id__candidates__candidate_id__crop_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "package_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Package Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/png": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "The integrity-checked mechanical crop."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "View the mechanical crop behind an untyped AI reading",
        "tags": [
          "confirmations"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/packages/{package_id}/layout-proposals/{crop_artifact_id}/crop": {
      "get": {
        "description": "Return a layout proposal's stored crop through the current package boundary.",
        "operationId": "layout_proposal_crop_api_v1_projects__project_id__packages__package_id__layout_proposals__crop_artifact_id__crop_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "package_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Package Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "crop_artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Crop Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "image/png": {}
            },
            "description": "The integrity-checked crop behind a proposed layout answer."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "View the crop behind a proposed closed layout answer",
        "tags": [
          "confirmations"
        ]
      }
    },
...
  }
}